Correctly handle zfs pools #413
