+++ src/detection/displayserver/linux/xlib.c
@@ -3,6 +3,7 @@
 #ifdef FF_HAVE_X11
 #include "common/library.h"
 #include "common/parsing.h"
+#include "util/stringUtils.h"
 #include <X11/Xlib.h>
 
 typedef struct X11PropertyData
