+++ src/detection/gpu/gpu_linux.c
@@ -6,6 +6,7 @@
 #include "common/properties.h"
 #include "common/parsing.h"
 #include "detection/temps/temps_linux.h"
+#include "util/stringUtils.h"
 #include <string.h>
 #include <unistd.h>
 #include <pci/pci.h>
