+++ src/detection/os/os_linux.c
@@ -1,6 +1,7 @@
 #include "os.h"
 #include "common/properties.h"
 #include "common/parsing.h"
+#include "util/stringUtils.h"
 
 #include <string.h>
 #include <stdlib.h>
