+++ src/detection/wmtheme/wmtheme.h
@@ -3,6 +3,8 @@
 #ifndef FASTFETCH_INCLUDED_detection_wmtheme
 #define FASTFETCH_INCLUDED_detection_wmtheme
 
+#include "fastfetch.h"
+
 bool ffDetectWmTheme(FFinstance* instance, FFstrbuf* themeOrError);
 
 #endif
