+++ src/detection/wmtheme/wmtheme_linux.c
@@ -1,10 +1,10 @@
-#include "fastfetch.h"
+#include "wmtheme.h"
 #include "common/properties.h"
 #include "common/parsing.h"
 #include "common/settings.h"
 #include "detection/gtk.h"
 #include "detection/displayserver/displayserver.h"
-#include "wmtheme.h"
+#include "util/stringUtils.h"
 
 static bool detectWMThemeFromConfigFile(FFinstance* instance, const char* configFile, const char* themeRegex, const char* defaultValue, FFstrbuf* themeOrError)
 {
