+++ src/fastfetch.c
@@ -1,9 +1,10 @@
 #include "fastfetch.h"
-#include "util/FFvaluestore.h"
 #include "common/printing.h"
 #include "common/parsing.h"
 #include "common/io/io.h"
 #include "common/time.h"
+#include "util/FFvaluestore.h"
+#include "util/stringUtils.h"
 
 #include <stdlib.h>
 #include <ctype.h>
