+++ src/util/FFstrbuf.h
@@ -18,8 +18,6 @@
     #define strcasestr StrStrIA
 #endif
 
-bool ffStrSet(const char* str);
-
 #define FASTFETCH_STRBUF_DEFAULT_ALLOC 32
 
 typedef struct FFstrbuf
